Separate custom matchers declarations into `.d.ts` file
