+++ __tests__/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "../tsconfig.json",
+  "compilerOptions": {
+    "skipLibCheck": false,
+    "noEmit": true
+  },
+  "files": ["_customMatchers.d.ts"]
+}
