+++ jest.config.cjs
@@ -1,7 +1,21 @@
+/* eslint-disable @typescript-eslint/no-var-requires */
+
+const { defaults: tsjPreset } = require('ts-jest/presets');
+
 /** @type {import('ts-jest').JestConfigWithTsJest} */
 module.exports = {
-  preset: 'ts-jest',
   testEnvironment: 'node',
   testMatch: ['**/__tests__/**/*.test.ts'],
+  setupFilesAfterEnv: ['./__tests__/customMatchers.ts'],
+  moduleFileExtensions: ['js', 'mjs', 'cjs', 'ts', 'd.ts', 'json'],
   verbose: true,
+  transform: {
+    '^.+\\.ts$': [
+      'ts-jest',
+      {
+        ...tsjPreset.transform,
+        tsconfig: './__tests__/tsconfig.json',
+      },
+    ],
+  },
 };
